Crate mirrord_intproxy

Modules

agent_conn

Implementation of proxy <-> agent connection through mpsc channels created in different mirrord crates.

background_tasks 🔒

Logic for managing background tasks in the internal proxy. The proxy utilizes multiple background tasks to split the code into more self-contained parts. Structs in this module aim to ease managing their state.

error

layer_conn 🔒

Implementation of layer <-> proxy connection through a TcpStream.

layer_initializer 🔒

main_tasks 🔒

ping_pong 🔒

Ping pong mechanism implementation on the internal proxy side. This mechanism exists to keep the proxy <-> agent connection alive when there are no requests from the layer.

proxies 🔒

Sub-proxies of the internal proxy. Each of these encapsulates logic for handling a group of related requests and exchanges messages only with the IntProxy.

remote_resources 🔒

request_queue 🔒

A request fifo used by the internal proxy to match agent responses with layer requests.

Structs

IntProxy

This struct contains logic for proxying between multiple layer instances and one agent. It maintains a singe agent connection.

TaskTxs 🔒

TaskSenders for main background tasks. See MainTaskId.