Module mirrord_intproxy::background_tasks

Logic for managing background tasks in the internal proxy. The proxy utilizes multiple background tasks to split the code into more self-contained parts. Structs in this module aim to ease managing their state.

Each background task implement the BackgroundTask trait, which specifies its properties and allows for managing groups of related tasks with one BackgroundTasks instance.

Structs

BackgroundTasks

A struct for managing groups of related BackgroundTasks. Tasks managed with a single instance of this struct must produce messages of the same type MOut and return errors convertible to Err.

MessageBus

A struct that is meant to be the only way the BackgroundTasks can communicate with their parents. It allows the tasks to send and receive messages.

TaskSender

A struct that can be used to send messages to a BackgroundTask registered in the BackgroundTasks struct. Dropping this sender will close the channel of messages consumed by the task (see MessageBus). This should trigger task exit.

Enums

TaskError

An error that can occur when executing a BackgroundTask.

TaskUpdate

An update received from a BackgroundTask registered in the BackgroundTasks struct.

Traits

BackgroundTask

Common trait for all background tasks in the internal proxy.