Trait mirrord_intproxy::background_tasks::BackgroundTask

pub trait BackgroundTask: Sized {
    type Error;
    type MessageIn;
    type MessageOut;

    // Required method
    fn run(
        self,
        message_bus: &mut MessageBus<Self>
    ) -> impl Future<Output = Result<(), Self::Error>> + Send;
}

Common trait for all background tasks in the internal proxy.

Required Associated Types

type Error

Type of errors that can occur during the execution.

type MessageIn

Type of messages consumed by the task.

type MessageOut

Type of messages produced by the task.

Required Methods

fn run( self, message_bus: &mut MessageBus<Self> ) -> impl Future<Output = Result<(), Self::Error>> + Send

Runs this tasks. The task can use the provided MessageBus for communication. When the MessageBus has no more messages to be consumed, the task should exit without errors.