Module mirrord_intproxy::protocol

Protocol used in communication between the layer and the internal proxy. This protocol does not have to be backwards compatible and can be changed freely, as the internal proxy and the layer are shipped together in a single binary.

Modules

macros 🔒

Utility macros for implementing traits on protocol types.

Structs

ConnMetadataRequest

A request for additional metadata for accepted connection. The layer should use this each time it accepts a connection on a socket that is listening for mirrored connections (PortSubscribe).

ConnMetadataResponse

A response to layer’s ConnMetadataRequest. Contains metadata useful for hooking getsockname and getpeername.

LayerId

Unique layer <-> proxy session identifier. New connection is established when the layer initializes or forks.

LocalMessage

A wrapper for messages sent through the layer <-> proxy connection.

OutgoingConnectRequest

A request to initiate a new outgoing connection.

OutgoingConnectResponse

A response to layer’s OutgoingConnectRequest.

PortSubscribe

A request to start proxying incoming connections.

PortUnsubscribe

A request to stop proxying incoming connections.

Enums

IncomingRequest

Requests related to incoming connections.

IncomingResponse

A response to layer’s IncomingRequest.

LayerToProxyMessage

Messages sent by the layer and handled by the internal proxy.

NetProtocol

Supported network protocols when intercepting outgoing connections.

NewSessionRequest

A layer’s request to start a new session with the internal proxy. Contains info about layer’s state. This should be the first message sent by the layer after opening a new connection to the internal proxy.

PortSubscription

Instructions for the internal proxy and the agent on how to execute port mirroring.

ProxyToLayerMessage

Messages sent by the internal proxy and handled by the layer.

Traits

IsLayerRequest

A helper trait for layer -> proxy requests.

IsLayerRequestWithResponse

A helper trait for layer -> proxy requests that require a response from the proxy. Not all layer requests require a response, e.g. CloseFileRequest.

Type Aliases

MessageId

An identifier for a message sent from the layer to the internal proxy. The layer uses this to match proxy responses with awaiting requests.