Trait mirrord_intproxy::protocol::IsLayerRequestWithResponse

pub trait IsLayerRequestWithResponse: IsLayerRequest {
    type Response: Sized;

    // Required methods
    fn wrap_response(response: Self::Response) -> ProxyToLayerMessage;
    fn try_unwrap_response(
        response: ProxyToLayerMessage
    ) -> Result<Self::Response, ProxyToLayerMessage>;
}

A helper trait for layer -> proxy requests that require a response from the proxy. Not all layer requests require a response, e.g. CloseFileRequest.

Note

Instead of this, we should ideally have something like IsProxyResponse trait. However, proxy <-> agent protocol uses the same response type for multiple requests. Translating agent responses into unique types would generate a lot of boilerplate code.

Required Associated Types

type Response: Sized

Type of response to this request.

Required Methods

fn wrap_response(response: Self::Response) -> ProxyToLayerMessage

Wraps the response so that it can be sent through the connection.

fn try_unwrap_response( response: ProxyToLayerMessage ) -> Result<Self::Response, ProxyToLayerMessage>

Tries to unwrap a response of valid type from the message. On error, returns the message as it was.