Module mirrord_intproxy::proxies::incoming

Handles the logic of the incoming feature.

Modules

http 🔒

http_interceptor 🔒

BackgroundTask used by Incoming to manage a single intercepted HTTP connection.

port_subscription_ext 🔒

Utilities for handling toggleable steal feature in IncomingProxy.

raw_interceptor 🔒

BackgroundTask used by Incoming to manage a single intercepted raw connection.

subscriptions 🔒

Structs

IncomingProxy

Handles logic and state of the incoming feature. Run as a BackgroundTask.

InterceptorHandle 🔒

InterceptorId

Id of a single interceptor task. Used to manage interceptor tasks with the BackgroundTasks struct.

MetadataStore 🔒

Enums

IncomingProxyError

Errors that can occur when handling the incoming feature.

IncomingProxyMessage

Messages consumed by IncomingProxy running as a BackgroundTask.

InterceptorError 🔒

Common type for errors of the RawInterceptor and the HttpInterceptor.

InterceptorMessageOut

Common type for messages produced by the RawInterceptor and the HttpInterceptor.