Trait mirrord_intproxy::proxies::incoming::port_subscription_ext::PortSubscriptionExt

pub trait PortSubscriptionExt {
    // Required methods
    fn port(&self) -> Port;
    fn agent_subscribe(&self) -> ClientMessage;
    fn wrap_agent_unsubscribe(&self) -> ClientMessage;
    fn wrap_agent_unsubscribe_connection(
        &self,
        connection_id: ConnectionId
    ) -> ClientMessage;
    fn wrap_response(
        &self,
        res: InterceptorMessageOut,
        connection_id: ConnectionId
    ) -> Option<ClientMessage>;
}

Trait for PortSubscription that handles differences in mirrord_protocol::tcp between the steal and the mirror flow. Allows to unify logic for both flows.

Required Methods

fn port(&self) -> Port

Returns the subscribed port.

fn agent_subscribe(&self) -> ClientMessage

Returns a subscribe request to be sent to the agent.

fn wrap_agent_unsubscribe(&self) -> ClientMessage

Returns an unsubscribe request to be sent to the agent.

fn wrap_agent_unsubscribe_connection( &self, connection_id: ConnectionId ) -> ClientMessage

Returns an unsubscribe connection request to be sent to the agent.

fn wrap_response( &self, res: InterceptorMessageOut, connection_id: ConnectionId ) -> Option<ClientMessage>

Returns a message to be sent to the agent in response to data coming from an interceptor. None means that the data should be discarded.