Trait mirrord_intproxy::proxies::outgoing::net_protocol_ext::NetProtocolExt

pub trait NetProtocolExt: Sized {
    // Required methods
    fn wrap_agent_write(
        self,
        connection_id: ConnectionId,
        bytes: Vec<u8>
    ) -> ClientMessage;
    fn wrap_agent_close(self, connection_id: ConnectionId) -> ClientMessage;
    fn wrap_agent_connect(self, remote_address: SocketAddress) -> ClientMessage;
    async fn prepare_socket(
        self,
        for_remote_address: SocketAddress
    ) -> Result<PreparedSocket>;
}

Trait for NetProtocol that handles differences in mirrord_protocol::outgoing between network protocols. Allows to unify logic.

Required Methods

fn wrap_agent_write( self, connection_id: ConnectionId, bytes: Vec<u8> ) -> ClientMessage

Creates a LayerWrite message and wraps it into the common ClientMessage type. The enum path used here depends on this protocol.

fn wrap_agent_close(self, connection_id: ConnectionId) -> ClientMessage

Creates a LayerClose message and wraps it into the common ClientMessage type. The enum path used here depends on this protocol.

fn wrap_agent_connect(self, remote_address: SocketAddress) -> ClientMessage

Creates a LayerConnect message and wraps it into the common ClientMessage type. The enum path used here depends on this protocol.

async fn prepare_socket( self, for_remote_address: SocketAddress ) -> Result<PreparedSocket>

Opens a new socket for intercepting a connection to the given remote address.