Struct mirrord_intproxy::proxies::outgoing::OutgoingProxy

pub struct OutgoingProxy {
    datagrams_reqs: RequestQueue,
    stream_reqs: RequestQueue,
    txs: HashMap<InterceptorId, TaskSender<Interceptor>>,
    background_tasks: BackgroundTasks<InterceptorId, Vec<u8>, Error>,
}

Handles logic and state of the outgoing feature. Run as a BackgroundTask.

Flow

Proxy receives an OutgoingConnectRequest from the layer and sends a corresponding LayerConnect to the agent.

Proxy receives a confirmation from the agent.

Proxy creates a new socket and starts a new outgoing Interceptor background task to manage it.

Proxy sends a confirmation to the layer.

The layer connects to the socket managed by the Interceptor task.

The proxy passes the data between the agent and the Interceptor task.

If the layer closes the connection, the Interceptor exits and the proxy notifies the agent. If the agent closes the connection, the proxy shuts down the Interceptor.

Fields
datagrams_reqs: RequestQueue

For OutgoingConnectRequests related to NetProtocol::Datagrams.

stream_reqs: RequestQueue

For OutgoingConnectRequests related to NetProtocol::Stream.

txs: HashMap<InterceptorId, TaskSender<Interceptor>>

TaskSenders for active Interceptor tasks.

background_tasks: BackgroundTasks<InterceptorId, Vec<u8>, Error>

For managing Interceptor tasks.

Implementations

impl OutgoingProxy

const CHANNEL_SIZE: usize = 512usize

Used when registering new Interceptor tasks in the BackgroundTasks struct.

fn queue(&mut self, protocol: NetProtocol) -> &mut RequestQueue

Retrieves correct RequestQueue for the given NetProtocol.

async fn handle_agent_read( &mut self, read: RemoteResult<DaemonRead>, protocol: NetProtocol ) -> Result<(), OutgoingProxyError>

Passes the data to the correct Interceptor task. Fails when the agent sends an error, because this error cannot be traced back to an exact connection.

async fn handle_connect_response( &mut self, connect: RemoteResult<DaemonConnect>, protocol: NetProtocol, message_bus: &mut MessageBus<Self> ) -> Result<(), OutgoingProxyError>

Handles agent’s response to a connection request. Prepares a local socket and registers a new Interceptor task for this connection. Replies to the layer’s request.

async fn handle_connect_request( &mut self, message_id: MessageId, session_id: LayerId, request: OutgoingConnectRequest, message_bus: &mut MessageBus<Self> )

Saves the layer’s request id and sends the connection request to the agent.