Module mirrord_intproxy::request_queue

A request fifo used by the internal proxy to match agent responses with layer requests.

The layer annotates each message with unique MessageId, which allows it to match responses from the internal proxy with awaiting requests. However, the agent does not use any mechanism like this. Instead, its components (e.g. file manager) handle requests of the same kind (e.g. FileRequests) sequentially. The internal proxy relies on this behavior and stores MessageIds of layer’s requests in multiple queues. Upon receiving a response from the agent, correct MessageId is taken from the right queue.

Additionaly, single internal proxy handles multiple layer instances (coming from forks). This fifo stores their LayerIds as well.

Structs

RequestQueue

A queue used to match agent responses with layer requests. A single queue can be used for multiple types of requests only if the agent preserves order between them.

RequestQueueEmpty

Erorr returned when the proxy attempts to retrieve MessageId and LayerId of a request corresponding to a response received from the agent, but the RequestQueue is empty. This error should never happen.