Struct mirrord_intproxy::IntProxy

pub struct IntProxy {
    pub(crate) any_connection_accepted: bool,
    pub(crate) background_tasks: BackgroundTasks<MainTaskId, ProxyMessage, IntProxyError>,
    pub(crate) task_txs: TaskTxs,
}

This struct contains logic for proxying between multiple layer instances and one agent. It maintains a singe agent connection.

Utilizes multiple BackgroundTasks to split logic of different mirrod features (e.g. file operations and incoming traffic).

Fields
any_connection_accepted: boolbackground_tasks: BackgroundTasks<MainTaskId, ProxyMessage, IntProxyError>task_txs: TaskTxs
Implementations

impl IntProxy

pub(crate) const CHANNEL_SIZE: usize = 512usize

Size of channels used to communicate with main tasks (see MainTaskId).

pub(crate) const PING_INTERVAL: Duration = _

How long can the agent connection remain silent.

pub async fn new( config: &LayerConfig, agent_connect_info: Option<AgentConnectInfo>, listener: TcpListener ) -> Result<Self, IntProxyError>

Initiates a new agent connection and creates a new IntProxy. The returned instance will accept connections from the layers using the given TcpListener.

pub fn new_with_connection( agent_conn: AgentConnection, listener: TcpListener ) -> Self

Creates a new IntProxy using existing AgentConnection. The returned instance will accept connections from the layers using the given TcpListener.

pub async fn run( self, first_timeout: Duration, idle_timeout: Duration ) -> Result<(), IntProxyError>

Runs main event loop of this proxy. Expects to accept the first layer connection within the given first_timeout. Exits after idle_timeout when there are no more layer connections.

pub(crate) async fn handle( &mut self, msg: ProxyMessage ) -> Result<(), IntProxyError>

Routes a ProxyMessage to the correct background task. ProxyMessage::NewLayer is handled here, as an exception.

pub(crate) async fn handle_task_update( &mut self, task_id: MainTaskId, update: TaskUpdate<ProxyMessage, IntProxyError> ) -> Result<(), IntProxyError>

Handles a TaskUpdate from one of the main tasks (see MainTaskId).

pub(crate) async fn handle_agent_message( &mut self, message: DaemonMessage ) -> Result<(), IntProxyError>

Routes most messages from the agent to the correct background task. Some messages are handled here.

pub(crate) async fn handle_layer_message( &self, message: FromLayer ) -> Result<(), IntProxyError>

Routes a message from the layer to the correct background task.